Fixed tests and organized into folders.
