+++ .github/workflows/check_ci.yaml
@@ -1,4 +1,4 @@
-name: Run Unit Tests
+name: Run Tests for CI Build
 
 on:
   push:
@@ -18,7 +18,7 @@ jobs:
        uses: ./.github/actions/prepare_poetry_env
 
      - name: Run pytest
-       run: poetry run pytest test/test_install_dependencies.py
+       run: poetry run pytest test/integration/test_install_dependencies.py
        env: # Set the secret as an env variable
           AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
           AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_ACCESS_KEY_SECRET }}
