+++ test/aws/__init__.py
@@ -1,0 +1,5 @@
+"""
+This package contains tests involving AWS resources.  In order to execute
+these tests you need an AWS account, a user with permissions in this account and
+an access key.
+"""
