+++ test/aws/cloudformation_validation.py
