+++ test/integration/__init__.py
@@ -1,0 +1,4 @@
+"""
+This package contains integration tests with longer runtime and maybe
+requiring additional resources.
+"""
