+++ test/integration/localstack_test.py
@@ -5,7 +5,7 @@ from exasol.ds.sandbox.lib.setup_ec2.cf_stack import CloudformationStack, \
     CloudformationStackContextManager
 from exasol.ds.sandbox.lib.setup_ec2.run_setup_ec2 import run_lifecycle_for_ec2
 from exasol.ds.sandbox.lib.tags import create_default_asset_tag
-from test.aws_local_stack_access import AwsLocalStackAccess
+from test.integration.aws_local_stack_access import AwsLocalStackAccess
 
 
 def test_ec2_lifecycle_with_local_stack(local_stack, default_asset_id, test_dummy_ami_id):
@@ -71,7 +71,7 @@ Resources:
           ToPort: 22
       Tags:
         - Key: "exasol:project"
-          Value: "ScriptLanguages"
+          Value: "DataScienceSandbox"
         - Key: "exasol:owner"
           Value: "test user"
 
@@ -90,7 +90,7 @@ Resources:
               VolumeSize: 100
           Tags:
             - Key: "exasol:project"
-              Value: "ScriptLanguages"
+              Value: "DataScienceSandbox"
             - Key: "exasol:owner"
               Value: "test"
