+++ test/unit/__init__.py
@@ -1,0 +1,4 @@
+"""
+This package contains unit tests requiring no additional setup or external
+resources.
+"""
