+++ test/unit/test_make_ami_public.py
